Refactor interdomain tests
